:art: FLOW-647 bind validaiton code segregated
