+++ packages/@cldcvr/flow-form-builder/src/components/f-form-builder/f-form-builder-constants.ts
@@ -1,0 +1,2 @@
+export const GROUP_FIELD_NAME_SEPARATOR = ".$";
+export const CLONNED_GROUP_NAME_SEPARATOR = "_$";
